+++ execution/consensus/misc/dao_params.go
@@ -17,7 +17,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package params
+package misc
 
 import (
 	"math/big"
